Move diagram utils to own package
