+++ pkg/diagrams/diagutil.go
@@ -1,4 +1,4 @@
-package main
+package diagrams
 
 import (
 	"bytes"
