+++ pkg/diagrams/diagutil_test.go
@@ -1,4 +1,4 @@
-package main
+package diagrams
 
 import (
 	"testing"
